requirements.txt
django==5.2.5
ipython==9.4.0
ruff==0.12.8
django-ninja==1.5.0
django-cors-headers==4.9.0
